requirements.txt
fastapi
uvicorn
sqlalchemy
httpx
aioredis
alembic
pydantic-settings
pydantic
passlib[bcrypt]
bcrypt<4.0.0
python-jose[cryptography]
asyncpg
email-validator
python-multipart
httpx
redis